pyproject.toml
[tool.poetry]
name = "vulnerable-restaurant"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^3.0.0"
mysqlclient = "^2.2.0"
python-dotenv = "^1.0.0"
oracledb = "^1.4.2"
psycopg2 = "^2.9.9"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
